jtagserv: Corrected error with singleton instance and logging
The issue I had been seeing about aji_read_device_chain() failure
has been traced back to two factors:

(1) The  singleton jtagservice
is not a singleton: jtagservice.c and jtagserv.c saw two different
instances. Hence, values sotred in jtagserv.c's jtagservice
was not passed to that in jtagservice.c.

(2) There is a problem with the logic of jtagservice_lock()
when it called jtagservice_is_locked(). The condition was
incorrect as it did not lock when it should.

Signed-off-by: Ooi, Cinly <cinly.ooi@intel.com>