HiFive2RevB : Incomplete access. ARM not working
Temporarily ARM is not working because I am just trying
to get UBII to access RISCV

RiscV access is established, it gets the IDCODE and issue
DTMCS instruction. Not yet working as in DR access has
yet to be established. Reason is the code for access is
too tailored for ARM

Signed-off-by: Ooi, Cinly <cinly.ooi@intel.com>