minjtagserv: Restore working copy
The tap order presented by openocd to minijtagserv driver
changed. Previously, in the configuration file TAPs
must be presented in the reverse order, and
taps will also be in reverse order in the tap list.

Now, both are in the normal order. This meant that
the verification of TAPs that has dependency on the
order the TAPs are presented are failing. To
resolve:
(1) Changed the order of TAP declaration in config file
(2) Remove depedency on checking that the correct TAP
is selected based on position of the TAP on the list,
used IDCODE instead

Signed-off-by: Ooi, Cinly <cinly.ooi@intel.com>