libjclient.so: Using dlopen()
Previously was doing direct call to mangled function
for Linux. Now we do it through dlopen()/dlsym().
This is in preparation for putting in Intel mandated
security hardening for DLL hijacking. [Yes, we are
using Linux, but the principles for DLL hijacking
still applies with libraries.so]

libjclient.so* must now be located at src
directory. This is to harmonize with Windows'
requirement for DLL placement with DLL hijacking
hardening in place.

What is interesting is c_aji_open_device_a(), i.e.
aji_open_device(AJI_CLAIM2) no longer work like
their windows counterpart. This suggests that
it is dynamic loading by application at runtime,
i.e. dlopen() or LoadLibrary(), that corrupts
the function.

Signed-off-by: Ooi, Cinly <cinly.ooi@intel.com>