NEW: Added aji_client driver
Implemented as minidriver. Minidriver interface is
deprecated and the patch for its removal is submitted
but not yet merged.

This is created to faciliate discussion on
to best take aji_client out to become a standard
openocd driver

src/jtag/aji_client is from release/21.1. Should
work fine, but not tested

Signed-off-by: Ooi, Cinly <cinly.ooi@intel.com>